parse expected_failure arithmetic and vector errors (#302)
* remove todo derive default on special att

* parse aritmetic error

* move set_or_insert_import to runtime

* implement tryfrom i32 for runtime error

* handle wasm trap in call test function

* link set_runtime_error

* stdlib test aritmetic error

* parse vector_error

* Update crates/move-parse-special-attributes/src/function_modifiers.rs


---------

+++ move-packages/move-stdlib/tests/u32_tests.move
@@ -52,9 +52,9 @@ fun test_divide_and_round_up() {
     integer_tests::test_divide_and_round_up!(MAX, CASES);
 }
 
-#[test, expected_failure(arithmetic_error, location = std::u8)]
+#[test, expected_failure(arithmetic_error, location = std::u32)]
 fun test_divide_and_round_up_error() {
-    1u8.divide_and_round_up(0);
+    1u32.divide_and_round_up(0);
 }
 
 #[test]
